Moved lock utilities to their own tree.
