+++ utilities/aws-utilities/pom.xml
@@ -47,9 +47,9 @@
         <coverage.line.ratio>0</coverage.line.ratio>
     </properties>
 
+    <!--
     <dependencies>
         <dependency>
-            <!-- for object mapper setup -->
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-json</artifactId>
         </dependency>
@@ -141,7 +141,6 @@
             </exclusions>
         </dependency>
         <dependency>
-            <!-- V1 API needed for the lock client -->
             <groupId>com.amazonaws</groupId>
             <artifactId>aws-java-sdk-dynamodb</artifactId>
             <scope>provided</scope>
@@ -169,7 +168,6 @@
             <version>14.2.0-SNAPSHOT</version>
         </dependency>
 
-        <!-- TEST DEPENDENCIES -->
         <dependency>
             <groupId>ch.qos.logback</groupId>
             <artifactId>logback-classic</artifactId>
@@ -181,4 +179,5 @@
             <scope>test</scope>
         </dependency>
     </dependencies>
+        -->
 </project>
