+++ utilities/lock-utilities/lock-dynamodb/src/lombok.config
@@ -1,0 +1,3 @@
+lombok.addLombokGeneratedAnnotation = true
+lombok.anyConstructor.addConstructorProperties=true
+config.stopbubbling=true
