+++ utilities/lock-utilities/lock-dynamodb/src/main/java/com/limemojito/aws/lock/DynamoDbLockService.java
@@ -20,6 +20,7 @@ package com.limemojito.aws.lock;
 import com.amazonaws.services.dynamodbv2.AcquireLockOptions;
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDBLockClient;
 import com.amazonaws.services.dynamodbv2.LockItem;
+import com.limemojito.lock.LockService;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
