+++ utilities/lock-utilities/lock-dynamodb/src/main/java/com/limemojito/aws/lock/LockConfig.java
