+++ utilities/lock-utilities/lock-dynamodb/src/test/resources/localstack-dynamodb/lock-table.json
