+++ utilities/lock-utilities/lock-postgres/src/main/java/com/limemojito/lock/postgres/PostgresLockConfig.java
@@ -1,0 +1,26 @@
+/*
+ * Copyright 2011-2024 Lime Mojito Pty Ltd
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ *
+ */
+
+package com.limemojito.lock.postgres;
+
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@ComponentScan(basePackageClasses = {PostgresLockConfig.class})
+public class PostgresLockConfig {
+}
