+++ utilities/lock-utilities/lock-postgres/src/main/java/com/limemojito/lock/postgres/PostgresLockService.java
@@ -15,29 +15,47 @@
  *
  */
 
-package com.limemojito.aws.lock;
+package com.limemojito.lock.postgres;
 
+import com.limemojito.lock.LockService;
 import lombok.Getter;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.dao.DataAccessException;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Optional;
 
 import static java.lang.String.format;
-import static org.springframework.transaction.annotation.Propagation.MANDATORY;
 
+/**
+ * The {@code PostgresLockService} class is an implementation of the {@link LockService} interface that provides methods for acquiring and managing distributed locks using PostgreSQL
+ * database.
+ * <p>
+ * The class utilizes the Postgres advisory lock functions to acquire and release locks. The locks are acquired within a transaction and are automatically released on transaction
+ * commit or rollback.
+ * </p>
+ *
+ * <p>
+ * To use {@code PostgresLockService}, an instance of {@link JdbcTemplate} is required to be passed in the constructor. The {@link JdbcTemplate} provides the necessary database access
+ * for acquiring and releasing locks.
+ * </p>
+ *
+ * <p>
+ * This class is thread-safe, and the methods for acquiring locks are synchronized to ensure proper concurrency control.
+ * </p>
+ *
+ * @see LockService
+ */
 @Service
 @Slf4j
+@RequiredArgsConstructor
 public class PostgresLockService implements LockService {
 
     private final JdbcTemplate template;
-
-    public PostgresLockService(JdbcTemplate template) {
-        this.template = template;
-    }
 
     /**
      * Lock with pg_try_advisory_xact_lock expecting to be in a transaction.   Unlock is with transaction commit or rollback.
@@ -48,7 +66,7 @@ public class PostgresLockService implements LockService {
      * @see PostgresDistributedLock
      */
     @Override
-    @Transactional(propagation = MANDATORY)
+    @Transactional(propagation = Propagation.MANDATORY)
     public synchronized Optional<DistributedLock> tryAcquire(String lockName) {
         final int lockNumber = name2Int(lockName);
         // If acquired this lock is released on transaction commit.
@@ -71,7 +89,7 @@ public class PostgresLockService implements LockService {
      * @see PostgresDistributedLock
      */
     @Override
-    @Transactional(propagation = MANDATORY)
+    @Transactional(propagation = Propagation.MANDATORY)
     public synchronized DistributedLock acquire(String lockName) {
         final int lockNumber = name2Int(lockName);
         template.execute(format("SELECT pg_advisory_xact_lock(%d)", lockNumber));
@@ -81,10 +99,16 @@ public class PostgresLockService implements LockService {
     /**
      * Unlock is with transaction commit or rollback.
      */
+    @Getter
     protected static class PostgresDistributedLock implements DistributedLock {
-        @Getter
         private final String name;
 
+        /**
+         * Represents a distributed lock in the PostgreSQL database.
+         * The lock is acquired when an instance of PostgresDistributedLock is created and released when the associated transaction is committed or rolled back.
+         *
+         * @param lockName the name of the lock
+         */
         protected PostgresDistributedLock(String lockName) {
             this.name = lockName;
             log.info("Acquired DB lock for {} -> {}", lockName, name2Int(lockName));
