+++ utilities/lock-utilities/pom.xml
@@ -21,23 +21,18 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>utilities</artifactId>
-    <description>Libraries for use in developing applications.</description>
-    <packaging>pom</packaging>
-
     <parent>
         <groupId>com.limemojito.oss.standards</groupId>
-        <artifactId>oss-maven-standards</artifactId>
+        <artifactId>utilities</artifactId>
         <version>14.2.0-SNAPSHOT</version>
-        <relativePath>../pom.xml</relativePath>
     </parent>
 
-    <modules>
-        <module>json-utilities</module>
-        <module>aws-utilities</module>
-        <module>lock-utilities</module>
-        <module>test-utilities</module>
-        <module>aws-lambda</module>
-    </modules>
+    <artifactId>lock-utilities</artifactId>
+    <packaging>pom</packaging>
 
+    <modules>
+        <module>lock-api</module>
+        <module>lock-dynamodb</module>
+        <module>lock-postgres</module>
+    </modules>
 </project>
